requirements.txt
moviepy
Pillow
google-auth
google-auth-oauthlib
google-api-python-client
tweepy
ollama
pyautogui
opencv-python
numpy
requests-oauthlib
whisper